Radio player engine: Initial implementation of RadioControlActor.
This is going to become the main entry point into the control package;
i.e. the main actor controlling the radio source to be played. For
now, only a minimum skeleton is available defining types required for
the implementation of helper service. This is going to be extended
together with more helper services becoming available.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>