RadioSourceStateActor: Changed the way to create instances.
Instead of an apply() function, the module now provides a Factory
trait together with a default implementation that allows creating the
behaviors for new instances. This approach makes it possible to test
the parameters passed when creating instances in test classes for
actors that depend on RadioSourceStateActor.

Therefore, this pattern can be used in general in the future for
actors that need to be created by other actors. The only drawback is
that the module can no longer be private, since it escapes its scope
via the Factory trait.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>