RadioControlActor: Extracted internal messages.
Added a new RadioControlProtocol module that is going to contain all
the messages exchanged internally between actors in the control
package which cannot be assigned to a single actor implementation.

By separating these messages, it becomes clearer which messages are
not part of an official actor protocol. Also, cyclic references
between the control actor and its children can be avoided.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>