Player engine: Added PlaybackContextFactoryActor class.
This is an actor implementation to manage the available playback
context factories. It is going to be used by PlaybackActor, which will
no longer be responsible for managing the factories itself.

Letting the factories be managed by a dedicated actor makes the code
cleaner. It also enables the sharing of factories between multiple
player actors.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>