PlaybackActor: A PlaybackContextFactoryActor is now used.
The creation of a PlaybackContext is done by delegating to this actor.
Currently, factories can still be added to the PlaybackActor; they are
then passed to the factory actor. This is going to be changed later.

Since the PlaybackContextFactoryActor now needs to be passed to the
PlaybackActor, changes were necessary at the PlayerFacadeActor and the
player implementations.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>