Radio player engine: Added a dedicated configuration class.
The radio player engine supports some more configuration properties on
top of the configuration for an audio player. For this purpose, the
RadioPlayerConfig class is now introduced. At this stage, in addition
to the embedded audio player configuration, some properties are
contained that are hard-coded so far in the source state service. The
set of properties is going to be extended later.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>