Radio player engine: Added ErrorStateActor module.
This module is going to host multiple actors and helper classes to
monitor radio sources whose playback failed for some reasons. The idea
is to check such sources in regular intervals, so that temporary
outages can be handled.

The module will be extended step by step with further functionality.
The initial implementation provides an internal helper class to
construct the actors required for testing playback of a source.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>