RadioSourceStateService: Added support for disabled sources.
Radio sources can now be (temporarily) disabled. They are then ignored
when searching for a replacement source. When the current source is
disabled a replacement source is played instead.

This mechanism is going to be used to disable radio sources whose
playback caused an error. In the future, sources can be disabled for
other reasons as well.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>