RadioPlayer: Introduced a new API.
This API is based on the functionality supported by the Radio control
actor and simpler than the existing one. The player is now responsible
for determining whether a source can be played and select a
replacement source if necessary.

For now, the old API is kept to not break client code. But after the
client has been adapted, it is going to be removed.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>