RadioSourceStateService: Improved handling of delays.
The calculation of the delay for the next source evaluation could fail
for dates in the far future, since the range of Java's Duration class
is limited. This has been fixed.

While at it, the maximumEvalDelay property of the radio player
configuration is now enforced when determining the next evaluation
time.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>