Radio player: Extracted the logic to update the status line.
Setting the status text, updating the playback time, and managing the
error indicator is now in the responsibility of a new
RadioStatusLineController class. This simplifies the main controller.

The logic to indicate an error state has been changed. The indicator
is now displayed only if the current (selected) source cannot be
played due to a playback error. Also, the logic to distinguish between
the current source and a replacement source could be simplified.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>