RadioEventConverterActor: Improved conversion of radio sources.
The RadioSources produced by the actor did not match the ones from the
configuration due to the manual conversion based on the URI of the
audio source. This caused problems in the event listeners.

To fix this, the actor now listens on radio events, too, and is thus
able to keep track on the currently played source. This one is then
used in the converted events.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>