ScheduledInvocationActor: API improvements.
Instead of having two different commands for typed and classic actor
invocations, there is now a new trait defining an actor invocation
with typed and classic implementations. Then a single
ActorInvocationCommand is sufficient that reference an instance of
this trait.

This simplifies the implementation and the usage of this actor. Also,
the concept of an actor invocation is useful on its own.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>