Radio player engine: Added MetadataConfig class.
The class defines a number of configuration settings related to
metadata of radio streams and exclusions. The idea is to disable radio
sources when specific patterns are found in the stream metadata. The
source can later be enabled again when the metadata changes.

The set of configuration options may be extended later if further use
cases are discovered.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>