Radio player engine: Added new RadioStreamBuilder class.
This class contains the functionality to create a RunnableGraph for
processing a radio stream defined by its URI. The graph allows access
to both the audio data and the metadata of the radio stream.

The corresponding functionality has been extracted from
RadioStreamActor. Having it available in a separate class, enables the
processing of radio streams for other purposes than radio playback.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>