Radio player engine: Added MetadataCheckActor module.
This module is going to provide functionality for checking metadata of
radio streams for configured exclusions. When an exclusion is matched
for the current radio source, the source is disabled, and the
sources's metadata is checked periodically whether the exclusion is
now over.

So far, an actor has been implemented which can receive metadata from
a radio stream. It is going to be used for checking disabled radio
sources.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>