Radio player engine: Converted MetadataConfig to a trait.
Also moved the function to query the metadata configuration of a
specific radio source from RadioSourceConfig to the new trait.

This allows for a stricter separation of concerns. For components only
interested in metadata configuration, it is now sufficient to have a
reference to a MetadataConfig instance; an additional
RadioSourceConfig object is no longer needed.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>