Radio player engine: Added RadioStreamManagerActor module.
This actor implementation is going to be responsible for the creation
of radio stream actors. The intention is that radio streams opened for
testing metadata should be reused when a successfully tested radio
source becomes active again. To achieve this, the manager actor keeps
actor references in a cache for a (rather short) time. When a new
stream actor for a radio source is requested it checks whether there
is a corresponding entry in the cache which can be reused. Otherwise,
a new instance is created.

The cache management is fully implemented, but support for closing
stream actors is still missing.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>