RadioStreamManagerActor: Stored latest metadata for cached sources.
When a stream actor from the cache is reused, it can happen that the
next update for metadata only happens after a while. In the meantime,
a player application can show no or stale data. To prevent this, it is
now possible to provide metadata when releasing a stream actor. When
it is then reactivated from the cache, a corresponding metadata event
is generated.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>