MetadataExclusionFinderService: Result is now a Future.
In the future, the service is going to have to calculate interval
queries, which is done asynchronously; hence, the result has been
changed to a Future. At the same time, a more complex response object
has been defined which also allows detecting stale results by
introducing a sequence number.

MetadataStateActor has been adapted to deal with the new asynchronous
nature of the service.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>