MetadataConfig: Added applicableAt property to MetadataExclusion.
With this new property it can be expressed, that the exclusion should
only be effective at specific points in time. This is useful for
instance for radio sources that do not reflect advertisements
specifically in their metadata, but only indicate that no title is
played. Combined with time intervals, it is then possible to conclude
that advertisements are played.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>