Player server: Extracted functions for network interfaces.
The function to fetch all relevant network interfaces is useful for
multiple clients. Therefore, it has been moved to a top-level
declaration in a new Network.scala file.

The file also contains an extension to obtain the local address of a
network interface. This is going to be used to determine the local
address a server has bound to.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>