Build: Fixed logging configuration.
In the slf4j dependencies, there were version conflicts between the
1.7.x and 2.x versions caused by the slf4j-simple module with a 2.x
version added as test dependency. This broke logging of the player
server application.

The fix was to remove this module and to rely on the log4-slf4j
implementation for logging also in tests. For this to work, it was
necessary to have a valid log4j2.xml on the classpath. This has been
achieved by adding a test log4j configuration to src/test/resources of
the `shared` project; this is effective, because `shared % Test` is
referenced by most other projects.

To configure logging for the player server, the easiest way
to have a valid logging configuration was to depend on the
log4jConfFragment. However, since this module declares asynchronous
loggers, it was also necessary to explicitly add a dependency to the
com.lmax.disruptor package.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>