ActorCreator: Added support for custom stop commands.
The function to create classic actors has been extended to support a
custom stop command. If specified, this command is sent to the actor
to stop it instead of a PoisonPill.

With this new feature, actors requiring a special stop handling can
now be created and managed via this trait.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>