EndpointRequestHandlerActor: The response can now be partly generated.
The response to be sent on a request can now be specified as a
template that can contain a variable for the current IP address. Since
the actor already knows to which local IP address it is bound, this
information can be reused for the bind address of the HTTP server.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>