Player server: The UI route can now be placed at the top level.
It is now possible to declare the UI route without a path prefix. That
means that all requests are handled by the route except for those
starting with the /api prefix. This simplifies serving content
consisting of multiple assets that reference each other via relative
paths. Since this content is not aware of the prefix used for the UI
route, relative paths cannot be resolved correctly.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>