ServiceFactorySpec: Extracted tests for routes.
Added a new RoutesSpec class to cover these tests. This class is going
to be extended by further tests for new routes to be defined.
Unfortunately, due to conflicting Scala 2 and 3 dependencies, it is
not possible to use the routes testkit from Akka HTTP; so, the tests
are run against a real server.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>