Server: Improved startup of the server.
The several steps to start the server are now done via Futures. This
makes sure that errors are correctly handled, and the actor system is
terminated.

Before this change, the process would just hang in case of an error to
start the server, e.g. if the configuration could not be loaded. This
was due to the actor system still running in the background.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>