Radio engine config: Added a README file.
The file gives some background information and describes the supported
configuration options. It makes sense to have this description at this
module than at a specific player implementation. The README file for
the radio player application is going to be changed to reference this
documentation (which partly has been copied from there).

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>