ServerConfigTestHelper: Enhanced creating of the ActorCreator.
It is now possible to specify a test kit for typed actors; then the
actor creation is delegated to this test kit. This seems to be
necessary, otherwise the creation of typed actors fails with an error
message that a top-level actor cannot be created from outside an actor
system with a custom user guardian actor.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>