PlayerServer: Enabled keep-alive of web socket connections.
In the application's configuration, the `periodic-keep-alive-max-idle`
was added. This makes Akka Http sending periodic PING frames to avoid
that the web socket connection gets closed when it is idle for too
long. This is necessary for this use case, since updates about the
radio playback state are not that frequent.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>