Build: Added a test3 project.
This project is used for the migration to Scala 3. It contains
duplicates of test helper classes defined by the `shared` and
`platform` projects. Modules migrated to Scala 3 cannot depend on the
original test classes, because this would cause a conflict between
Scala 2 and 3 modules of dependencies. Therefore, it is necessary to
work with temporary copies.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>