Archive common: Reworked parsing of medium info files.
The XML files with medium information are now parsed using Commons
Configuration instead of the Scala XML module. Commons Configuration
is already used as a dependency, and it is well-suited for this
purpose. So, there is no need to introduce another dependency.

The Scala XML module also caused problems when trying to migrate the
archive modules to Scala 3.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>