ValidationController: Added a type parameter.
The constructor was using a wildcard as type parameter for the managed
`ValidationFlow`. This is not allowed in Scala 3 and causes an
"unreducible application of higher-kinded type" error. The work-around
is to make the class generic in the result type.

See: https://docs.scala-lang.org/scala3/guides/migration/incompat-other-changes.html#wildcard-type-argument

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>