Build: Set global Scala version to 3.
All modules are now built using Scala 3, including the OSGi images.
The module-specific overrides of Scala version and options have been
removed.

Unfortunately, it is not possible to drop the cross-build declarations
and use only Scala 3 versions of the dependencies. This does not work
for the OSGi images because of inconsistent imports in the bundle
versions for Scala 3.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>