RadioControlActor: Extended the current playback state.
The playback state now includes information about the currently played
title. This is relevant information for clients to be displayed in the
UI. Although it is possible to obtain this information via radio
events, there is no guarantee when the next metadata event is
received, since this depends on the radio source.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>