LineWriterStage: The line is now created dynamically.
Changed the management of the line, so that the stage is now
responsible for its creation on receiving the header of the audio
stream. Based on the format in the stream, a compatible
`SourceDataLine` can be created. The line is now always closed on
completion of the stage.

To make this possible, `LineWriterStage` now operates on the
`AudioData` type introduced by `AudioEncodingStage`.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>