Build: Added a dedicated project for the audio player shell.
Moved the `AudioPlayerShell` class into a separate project. This is
needed, since it requires a direct dependency to the Mp3ContextFactory
project (which is not the case for the audio player UI project that
hosted this class before).

The new `AudioStreamFactory` implementations have been integrated, so
that playback of MP3 files should be possible.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>