Player engine: Reworked AudioEncodingStage implementation.
The previous implementation had some severe flaws leading to an
uncontrolled memory consumption and interrupted playback.

Now the size of the in-memory buffer can be specified, and data from
the source is only fetched until the buffer is full. The whole logic
of pulling from upstream and pushing to downstream has been changed to
make sure that only the minimum required data passes the stream at the
correct point of time.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>