Player engine: Initial implementation of AudioStreamPlayerStage.
This stage can execute a whole audio stream for an audio source. It
can be integrated in a stream representing a playlist. Each item in
the playlist stream is mapped to an audio source, which gets played by
the stage.

Currently, only the minimum functionality has been implemented to run
an audio stream.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>