AudioStreamPlayerStage: Moved the kill switch in the audio stream.
In order to be effective immediately, the kill switch to skip
the audio stream needs to be placed after the encoding stage.
Otherwise, playback continues until the buffer of the encoding stage
is exhausted.

Unfortunately, it does not seem to be possible with reasonable effort
to write a unit test that checks the placement of the kill switch.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>