Player engine: Started work on BufferedPlaylistSource.
This is a module that allows buffering of audio sources on the local
file system before they get played. It offers the functionality which
had been provided by multiple actors originally, by making use of
streams.

In the current state of the implementation, multiple audio sources can
be combined and written into a buffer file.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>