BufferedPlaylistSource: Changed type of flow to fill the buffer.
The flow stage to copy sources into the buffer now keeps track on the
sources and their sizes that have been written into a buffer file.
When a buffer file is completed, a data object with this information
is emitted downstream. Via this mechanism, other components get the
information about which data is available in the buffer.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>