BufferedPlaylistSource: Improved error handling.
Errors when resolving a new audio source or from the source while
filling it into the buffer are now handled.

The bridge actor used by `FillBufferFlowStage` is now also stopped
when stream processing is complete. (Otherwise all test cases except
for the first one would fail because the name of the actor is already
in use.)

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>