BufferedPlaylistSource: Started with reading from the buffer.
Introduced a new internal `ReadBufferFlowStage` class that reads the
original audio sources again from the temporary buffer files. Also
provided an initial implementation of the `apply()` function, so that
the buffered source can now be tested as a whole.

In the current implementation, only a single source can be read from a
single buffer file.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>