BufferedPlaylistSource: Small overlaps in buffer files are handled.
Sources that end in the beginning of a new buffer file (at a position
smaller than the default read chunk size) were not be handled
correctly. This has now been fixed by involving the bridge actor into
updates of source limits. Pending requests for data chunks are
invalidated and have to be resent with the updated limit.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>