BufferedPlaylistSource: Handled some corner cases when reading.
On the read side, there can be a number of subtle race conditions
between a completely read buffer file and a source that has been
processed; especially if both happens almost at the same time. To deal
with those, the bridge actor has been extended to coordinate the
update of the size of sources that span multiple buffer files.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>