BufferedPlaylistSource: Implemented cleanup of buffer folder.
The buffer files are now deleted after they have been read. There is
also an extra cleanup step at the end to make sure that all files are
deleted, even if processing was canceled in the middle.

The test whether the file buffer is actually used had to be rewritten;
it obviously never worked as expected.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>