Player Server: Configured a log file.
This is especially helpful when starting the server on a device on the
network to diagnose problems. In a concrete setup on a Raspberry Pi,
it happens from time to time that the network discovery does not work.
With the log file, it is hopefully easier to discover the problem.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>