Player Server: Made `EndpointRequestHandlerActor` more robust.
The actor now checks first whether network interfaces are available it
can bind against. If this is not the case, it retries with an
increasing delay. Without this change, the actor could not handle
requests when it was started before a network interface (such as
Wi-Fi) was available, even if it was enabled at a later point in time.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>