BufferedPlaylistSource: Fixed problems with upstream finished.
The existing implementation was expecting that the upstream finished
notification arrived while data is still written into a buffer file.
This is not necessarily the case, for instance, if a queue is used as
source that can be closed at any time. This can now be handled; also,
the implementation can now deal with an empty playlist.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>