Build: Updated Apache Pekko dependencies.
The goal was to update to Apache Pekko 1.1.0. This is, however,
currently not possible due to incorrect OSGi metadata in most Apache
Pekko artifacts. Therefore, a locally built patch version is currently
referenced, in the hope that the fix makes it into the next official
release.

In the new Pekko version, there have been major changes wrt logging.
To deal with this, logging dependencies have been adapted. However,
the solution is currently not fully functional in OSGi; here no logs
from the Actor System are written. This still has to be investigated.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>