ActorSystem: Reworked the `Activator` class.
After switching to Pekko 1.1.x, the logs of the actor system no longer
showed up. Reason was a race condition in the startup order of
bundles: When the actor system bundle started the logger (actually
provided by the SpiFly dynamic weaving bundle) was not yet available.

The `Activator` class now deals with this scenario by waiting until
the SpiFly bundle is started. This is a work-around; there may be a
cleaner solution. To be investigated later.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>