osgi: Changed handling of service loaders.
Instead of using the static Aries SpiFly bundle to enable the service
loader required for obtaining an audio stream, this is now handled by
the dynamic weaving bundle. The dynamic bundle is needed anyway to
make logging work; tests have shown that this also handles the
services for the javax.sound system.

This simplifies the build process, since the SpiFly plugin is no
longer required.

Also updated the OSGi plugin.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>