Archive commons: Reworked `MetaDataParser` class.
The companion object now provides a function for parsing sources in
the expected `mdt` format. The parsing is based on the
`JsonStreamParser` module. The existing functionality is still
present, since it is used by clients. It is going to be removed after
the clients have been changed accordingly.

The new parsing function uses a different approach to error handling:
It is not lenient, but throws errors for invalid input data. This
should be fine, since the data to be parsed is also generated by code
in this repository and should therefore be correct.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>