Radio engine: Started implementation of `AttachableSink`.
This is a special implementation of a `Sink` that allows to
dynamically attach a consumer to a running stream. This is needed when
checking for metadata exclusions and then reusing a stream for audio
playback.

Since the implementation is going to be non-trivial, it is broken down
into several steps. So far, only basic stream processing logic has
been implemented without support for attaching a consumer.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>