AttachableSink: Reworked stopping of the control actor.
The actor is now responsible for stopping itself when it processes a
message that indicates stream completion or failure. This prevents
potential race conditions, especially that it is stopped before the
last element is passed to the source.

The state can now store multiple elements to be passed to the consumer
source. This is necessary for messages indicating the end of the
stream which can be sent by the sink before it receives the
confirmation about a processed message.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>