Radio engine: Changed `RadioStreamBuilder.buildRadioStream()`.
The function now expects a new parameters object containing the
required information. This is more readable and flexible than the
original large parameters list.

The parameters object no longer contains the full player
configuration. This was used before only to compute the size of the
playback buffer. This parameter is now passed in explicitly.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>