Radio engine: Fixed a problem with `MetadataExtractionStage`.
The stage was eagerly pushing metadata downstream. This was only
working if both sinks for audio data and metadata were directly
connected. When using an `AttachableSink`, an
`IllegalStateException` could occur because elements were pushed
before receiving a pull signal. This could be fixed by using the
`emit` API instead of the `push` API.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>