RadioStreamHandle: Changed the way to create instances.
Instead of the factory function in the companion object, now a factory
trait has been introduced, together with a default implementation.
This was mainly done to improve testability; so that defined handles
can be created by injecting a mock factory.

The factory function now also supports a parameter for the stream name
that is used to derive the names for attachable sinks. That way,
multiple streams can be created in parallel whose control actors must
have unique names.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>