Radio engine: Added `PlaybackStateActorNew` class.
This is an actor implementation copied from `PlaybackStateActor` that
uses a `RadioStreamPlaybackActor` to control the playback of the
current radio source. It is going to replace the original playback
state actor later.

The reason for introducing this class and adding temporary duplication
is that the change from the actor-based radio player engine to the new
stream-based one can then be done in smaller steps - instead of
updating the whole chain of callers in a big bang.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>