AudioEncodingStage: Added parameter validation.
The stage now checks whether the stream limit from the audio factory
does not exceed the memory buffer size. This would lead to a blocked
stream. If this error condition is detected, the stage fails the
stream with a corresponding error message.

Also added some log messages at interesting points of the stage's
life-cycle.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>