RadioStreamPlaybackActor: Radio stream names are now unique.
So far, a constant name was used for radio streams. This caused
problems when switching to another source, because for a short time
both the old and the new stream were active. Therefore, multiple
control actors for the stream handle were created with the same names.

This has now been fixed by appending the current sequence number of
the playback actor state to the stream name.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>