RadioSource: Added a property to query the URI with extension.
In order to correctly resolve the audio stream, it is required that
the URI of the radio stream has the correct file extension. So far,
this check was implemented in `RadioStreamPlaybackActor`. In the
future, this has to be done at other places as well. Therefore, the
functionality was moved to the `RadioSource` class.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>