ErrorStateActor: Reworked checks for radio playback.
The check whether a radio source can be played is now done via the new
radio player engine, which means via the stream handle manager actor
and stream handles. This simplifies the code, and it is no longer
necessary to create classic actors for managing the radio source and
playing audio data.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>