BufferedPlaylistSource: Full rework of processing logic.
To avoid race conditions (and fix the flaky) tests, more
responsibility for stream control has been shifted towards the bridge
actor. It now keeps track of the number of bytes processed by the
sources. Thus, it can determine when they are complete, handle
skipping logic and deal with many corner cases.

Note: With the changes that have been implemented currently, the tests
are green; but there is still some unneeded code that can be
simplified or removed.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>