MediaMetadata: Fixed a `ClassCastException` on serialization.
After recent updates of dependencies, suddenly a `ClassCastException`
occurs when sending `MediaMetadata` objects over the network. The
reason seems to be that numeric values are now handled differently,
and the file size, which is declared as `Long`, is serialized as an
`Int`. To fix this, the property was changed to an `Int`.

This might not be the best solution, since the file size (in theory)
could exceed the maximum value of an `Int`. However, with the
processed audio files, this is not likely. Also, it is planned to
rework the media archive to be based on an HTTP service; this would
use different means for serialization.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>