MetadataManagerActor: Removed metadata union actor reference.
The actor no longer directly sends messages to the metadata union
manager actor. Instead, only the metadata processing listener is
invoked. The startup of the local media archive has been reworked to
pass a correct listener actor here which delegates to the metadata
union actor, so that the behavior stays the same.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>