MediaMetadata: Added a `checksum` property.
This property is going to be used as a way to uniquely identify a
media file. It also allows finding duplicate files in a
straight-forward way.

The `size` property of the metadata is no longer optional, since it is
expected that it can always be determined during a scan operation.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>