id3 extract: Transformed `ID3v2ExtractorStage` to a sink.
This is more convenient for the use case at hand. The idea is to have
a processing stream with multiple sinks for different purposes. The
materialized values of the sinks can then be combined to produce the
final extraction result.

While at it, improved error handling by dealing with upstream errors.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>