id3Extract: Removed obsolete classes.
Almost all classes from the `processor` package are now obsolete,
since extraction of ID3 tags is now handled by the classes in the
`stream` package. Also, the `TailBuffer` class is no longer needed.

The `ID3v2ProcessingStage` is an exception; it is still needed to
strip ID3 tags when downloading media files from the archive.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>