Build: Extended `DefaultExcludedModules` sequence.
Some of the dependencies which are valid OSGi bundles cannot be
directly deployed in an OSGi environment using the Scala 3 library,
because they import Scala packages with a version 2.13 constraint.
Such bundles have to be excluded from the OSGi image generation; they
are replaced by manually created bundles with compatible metadata.

Moved the `DefaultExcludedModules` field to the `OSGi` object where
all OSGi-related settings should be located.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>