LoadPlaylistActor: `JsonStreamParser` is used for the position.
The file with the playlist position is now loaded via a new function
offered by `JsonStreamParser`.

The handling of the result of the asynchronous operation has been made
more robust by doing this in `receive` as reaction on a new message
instead of accessing the actor context from a future.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>