shared: Changed `MediumInfo` class.
The properties describing the medium and that need to be read from a
description file have been moved to a dedicated `MediumDescription`
class. This is going to simplify serialization code. While this change
does not affect code that only reads properties (due to the usage of
Scala's `export` feature), the creation of instances had to be
adapted.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>