archiveServer: Improved `ArchiveContentMetadataProcessingListener`.
The actor now provides a `Factory` trait to simplify the creation of
instances and make this easier testable.

An instance now stops itself on receiving a `ProcessingCompleted`
event. The idea is that on startup of the server the local media
archive is processed. When this is done, the listener is no longer
needed.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>