archiveServer: Added `ArchiveServerConfig` class.
The data class represents the configuration of the archive server.
This includes the configuration of the media archives to load. There
are also properties to configure the server itself. For now, there is
only a property for the port of the HTTP server; this may be extended
in the future.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>