archive: Added `ArchiveTocWriter` module.
This module is going to replace the `ArchiveTocWriterActor` class. It
is able to create a new "table of content" file for an archive or
update an existing one. The file contains information about the
last-modified time of single media. It is only written if there are
changes.

This is a preparation to have a local caching mechanism for HTTP
archives. This requires a mechanism to detect when there have been
changes on specific media.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>