mediaArchive: Introduced a new `cloudAccess` project.
This project is going to hold functionality to integrate file systems
from the `CloudFiles` project with media archives. It is going to be
used by the projects implementing HTTP archives and by an upcoming
archive server project for cloud archives.

So far, only a data class to hold a file system with some additional
info has been moved. It has been renamed from `HttpArchiveFileSystem`
to `CloudArchiveFileSystem`, since "cloud archive" is going to become
the new term for such kind of archives.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>